+++ lib/docket/checkpoint.ex
@@ -17,6 +17,7 @@ defmodule Docket.Checkpoint do
           :run_initialized
           | :step_committed
           | :retry_scheduled
+          | :detach_scheduled
           | :interrupt_requested
           | :interrupt_resolved
           | :run_completed
@@ -36,6 +37,7 @@ defmodule Docket.Checkpoint do
     :run_initialized,
     :step_committed,
     :retry_scheduled,
+    :detach_scheduled,
     :interrupt_requested,
     :interrupt_resolved,
     :run_completed,